requirements.txt
# Core Dependencies
Flask==2.0.1
Werkzeug==2.0.3
gunicorn==20.1.0

# Google Cloud Dependencies
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.14.0
google-cloud-translate>=3.12.0
google-cloud-vision>=3.7.0
google-cloud-speech>=2.21.0
firebase-admin>=6.4.0

# Authentication & Session Management
PyJWT>=2.8.0
Flask-Session>=0.5.0
cryptography>=42.0.0

# Image Processing
Pillow>=10.0.0

# Development Dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0

# Production Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
